Various changes
Updated lighting of the sample scene arena.
Created a parent class for characters.
Created a player and enemy class from the character parent.
Created a player script that will be the basis of the player's scripting.
Created a skeleton script that will be the basis of the skeleton enemy scripting.
(Scripts created are just tests so far to verify methods are created/called correctly)